+++ src/storage/file_manager.cpp
@@ -32,12 +32,12 @@ FileManager::FileManager(std::string dir) : dir_(std::move(dir)) {}
 
 FileManager::~FileManager() { close_all(); }
 
-std::string FileManager::table_file_path(uint32_t table_id) const
+std::string FileManager::table_file_path(uint64_t table_id) const
 {
     return dir_ + "/t_" + std::to_string(table_id) + ".dat";
 }
 
-int FileManager::fd_for(uint32_t table_id)
+int FileManager::fd_for(uint64_t table_id)
 {
     auto it = fds_.find(table_id);
     if (it != fds_.end()) {
@@ -52,7 +52,7 @@ int FileManager::fd_for(uint32_t table_id)
     return fd;
 }
 
-uint64_t FileManager::file_size(uint32_t table_id) const
+uint64_t FileManager::file_size(uint64_t table_id) const
 {
     auto it = fds_.find(table_id);
     if (it == fds_.end()) {
@@ -69,7 +69,7 @@ uint64_t FileManager::file_size(uint32_t table_id) const
     return static_cast<uint64_t>(st.st_size);
 }
 
-void FileManager::create_table_file(uint32_t table_id)
+void FileManager::create_table_file(uint64_t table_id)
 {
     const int fd = fd_for(table_id);
     if (::ftruncate(fd, static_cast<off_t>(PAGE_SIZE)) != 0) {
@@ -77,7 +77,7 @@ void FileManager::create_table_file(uint32_t table_id)
     }
 }
 
-void FileManager::remove_table_file(uint32_t table_id)
+void FileManager::remove_table_file(uint64_t table_id)
 {
     if (fds_.count(table_id) != 0) {
         ::close(fds_[table_id]);
@@ -89,18 +89,18 @@ void FileManager::remove_table_file(uint32_t table_id)
     }
 }
 
-bool FileManager::table_file_exists(uint32_t table_id) const
+bool FileManager::table_file_exists(uint64_t table_id) const
 {
     struct stat st {};
     return ::stat(table_file_path(table_id).c_str(), &st) == 0;
 }
 
-uint32_t FileManager::page_count(uint32_t table_id) const
+uint32_t FileManager::page_count(uint64_t table_id) const
 {
     return static_cast<uint32_t>(file_size(table_id) / PAGE_SIZE);
 }
 
-void FileManager::read_page(uint32_t table_id, uint32_t page_no, char* out)
+void FileManager::read_page(uint64_t table_id, uint32_t page_no, char* out)
 {
     const int fd = fd_for(table_id);
     const off_t off = static_cast<off_t>(static_cast<uint64_t>(page_no) * PAGE_SIZE);
@@ -113,7 +113,7 @@ void FileManager::read_page(uint32_t table_id, uint32_t page_no, char* out)
     }
 }
 
-void FileManager::write_page(uint32_t table_id, uint32_t page_no, const char* data)
+void FileManager::write_page(uint64_t table_id, uint32_t page_no, const char* data)
 {
     const int fd = fd_for(table_id);
     const off_t off = static_cast<off_t>(static_cast<uint64_t>(page_no) * PAGE_SIZE);
@@ -126,14 +126,14 @@ void FileManager::write_page(uint32_t table_id, uint32_t page_no, const char* da
     }
 }
 
-uint32_t FileManager::append_page(uint32_t table_id, const char* data)
+uint32_t FileManager::append_page(uint64_t table_id, const char* data)
 {
     const uint32_t new_page_no = page_count(table_id);
     write_page(table_id, new_page_no, data);
     return new_page_no;
 }
 
-void FileManager::flush(uint32_t table_id)
+void FileManager::flush(uint64_t table_id)
 {
     auto it = fds_.find(table_id);
     if (it == fds_.end()) {
