+++ src/storage/file_manager.h
@@ -19,33 +19,33 @@ public:
     FileManager& operator=(const FileManager&) = delete;
 
     std::string dir() const { return dir_; }
-    std::string table_file_path(uint32_t table_id) const;
+    std::string table_file_path(uint64_t table_id) const;
 
     // 建新表文件, 初始只含文件头页
-    void create_table_file(uint32_t table_id);
+    void create_table_file(uint64_t table_id);
     // 删除表文件(只删文件, 不动目录)
-    void remove_table_file(uint32_t table_id);
-    bool table_file_exists(uint32_t table_id) const;
+    void remove_table_file(uint64_t table_id);
+    bool table_file_exists(uint64_t table_id) const;
 
-    uint32_t page_count(uint32_t table_id) const;
-    void read_page(uint32_t table_id, uint32_t page_no, char* out);
-    void write_page(uint32_t table_id, uint32_t page_no, const char* data);
+    uint32_t page_count(uint64_t table_id) const;
+    void read_page(uint64_t table_id, uint32_t page_no, char* out);
+    void write_page(uint64_t table_id, uint32_t page_no, const char* data);
     // 在文件末尾追加一页, 返回新页页号
-    uint32_t append_page(uint32_t table_id, const char* data);
+    uint32_t append_page(uint64_t table_id, const char* data);
 
     // fsync 指定表文件
-    void flush(uint32_t table_id);
+    void flush(uint64_t table_id);
     void flush_all();
 
     // 关闭并释放所有文件描述符
     void close_all();
 
 private:
-    int fd_for(uint32_t table_id);
-    uint64_t file_size(uint32_t table_id) const;
+    int fd_for(uint64_t table_id);
+    uint64_t file_size(uint64_t table_id) const;
 
     std::string dir_;
-    std::unordered_map<uint32_t, int> fds_;
+    std::unordered_map<uint64_t, int> fds_;
 };
 
 }  // namespace st
