+++ src/storage/storage.h
@@ -58,7 +58,7 @@ public:
     // 刷盘并关闭
     void close();
 
-    uint32_t create_table(const std::string& name, const std::vector<ColumnSpec>& cols);
+    uint64_t create_table(const std::string& name, const std::vector<ColumnSpec>& cols);
     // 删表, 保留段表拒绝删除
     void drop_table(const std::string& name);
     RowRef insert(const std::string& table, const std::vector<Value>& values);
@@ -76,24 +76,24 @@ public:
 private:
     friend class Scanner;
     // 建首个数据页(页号 1)并链到文件头页, 返回新页号
-    uint32_t link_header_to_first_data_page(uint32_t table_id);
+    uint32_t link_header_to_first_data_page(uint64_t table_id);
     // 建表公共路径(须持锁): 校验后按指定 id 建数据文件、写元数据行
-    uint32_t create_table_impl(const std::string& name, const std::vector<ColumnSpec>& cols,
-                               uint32_t tid);
+    uint64_t create_table_impl(const std::string& name, const std::vector<ColumnSpec>& cols,
+                               uint64_t tid);
     // 物理建表(须持锁): 建数据文件并初始化落盘文件头页
-    void init_table_file(uint32_t tid);
+    void init_table_file(uint64_t tid);
     // 插行公共路径(须持锁): 校验编码后追加, 用户插行与元数据表引导共用
-    RowRef insert_impl(uint32_t table_id, const std::vector<ColumnSpec>& cols,
+    RowRef insert_impl(uint64_t table_id, const std::vector<ColumnSpec>& cols,
                        const std::vector<Value>& values);
     // 写入指定表的元数据行(须持锁): db_table 一行, db_column 每列一行, 引导与建表共用
-    void write_meta_rows(uint32_t tid, const std::string& name,
+    void write_meta_rows(uint64_t tid, const std::string& name,
                          const std::vector<ColumnSpec>& cols);
     // 引导元数据表: 直接建数据文件并写入自描述行, 不经过元数据表查找
     void bootstrap_meta_tables();
     // 删除指定表的元数据行(须持锁): 按 table_id 匹配 db_table/db_column
-    void delete_meta_rows(uint32_t tid);
+    void delete_meta_rows(uint64_t tid);
     // 读取指定表全部存活行(须持锁): 沿页链解码, 行损坏当场报错
-    std::vector<std::vector<Value>> read_rows(uint32_t table_id,
+    std::vector<std::vector<Value>> read_rows(uint64_t table_id,
                                               const std::vector<ColumnSpec>& cols);
     // 按表名查元数据(须持锁): db_table 定位 id, db_column 收集列并按 ordinal 排序,
     // 表不存在或元数据行非法当场报错
@@ -101,14 +101,14 @@ private:
     // 表名是否已存在(须持锁): 全扫 db_table 匹配
     bool has_table_name(const std::string& name);
     // table_id 是否已存在(须持锁): 全扫 db_table 匹配
-    bool table_id_exists(uint32_t table_id);
+    bool table_id_exists(uint64_t table_id);
     // 用户段分配(须持锁): max(当前最大表 id + 1, kFirstUserTableId)
-    uint32_t alloc_table_id();
+    uint64_t alloc_table_id();
 
     std::string dir_;
     FileManager files_;
     BufferPool pool_;
-    std::unordered_map<uint32_t, uint32_t> tail_pages_;  // 表 -> 最高页号(含仅存内存的页)
+    std::unordered_map<uint64_t, uint32_t> tail_pages_;  // 表 -> 最高页号(含仅存内存的页)
     std::mutex mutex_;   // 序列化所有操作(并发演化见设计文档 §10)
     bool open_ = false;
 };
